Rename. To many components named Label.
